requirements.txt
flask
mysql-connector-py